Consoldiated and factored out functionality for managing and responding to logon requests. More work needed.

+++ Securifi Cloud/UI/SFISceneViewController.m
@@ -606,7 +606,7 @@ static NSString *simpleTableIdentifier = @"SensorCell";
 }
 
 -(void)didRotateFromInterfaceOrientation:(UIInterfaceOrientation) fromInterfaceOrientation {
-    NSLog(@"Rotation %ld", fromInterfaceOrientation);
+    NSLog(@"Rotation %ld", (long) fromInterfaceOrientation);
     [self.tableView reloadData];
 }
